Cleanup and documentation
